Bugfixes, disabled a couple of features. 2.6.6.

+++ floatingimage/src/dk/nindroid/rss/ImageCache.java
@@ -23,14 +23,12 @@ import android.os.Environment;
 import android.util.Log;
 import dk.nindroid.rss.data.FileDateReverseComparator;
 import dk.nindroid.rss.data.ImageReference;
-import dk.nindroid.rss.settings.Settings;
 
-public class ImageCache implements Runnable {
+public class ImageCache {
 	TextureBank bank;
 	private final String 	mExploreFolder;
 	private final String 	mExploreInfoFolder;
-	private boolean			mActive = true;
-	List<File>				mExploreFiles;
+	List<File>				mFiles;
 	Map<String, Integer>	mCached;
 	Random 					mRand;
 	File 					mExplore;
@@ -50,7 +48,7 @@ public class ImageCache implements Runnable {
 		mExplore.mkdirs(); // Make dir if not exists
 		File[] exploreInfoArray = mExploreInfo.listFiles();
 		if(exploreInfoArray == null) return;
-		mExploreFiles = new ArrayList<File>(exploreInfoArray.length);
+		mFiles = new ArrayList<File>(exploreInfoArray.length);
 		mCached = new HashMap<String, Integer>(exploreInfoArray.length);
 		Log.v("Floating Image", exploreInfoArray.length + " files in cache.");
 		for(int i = 0; i < exploreInfoArray.length; ++i){
@@ -62,14 +60,7 @@ public class ImageCache implements Runnable {
 		}
 	}
 	
-	public void pause(){
-		mActive = false;
-	}
-	
-	public void resume(){
-		mActive = true;
-	}
-	
+	/*
 	public void run(){
 		if(mExploreFiles == null) return;
 		while(true){
@@ -88,7 +79,7 @@ public class ImageCache implements Runnable {
 			}
 		}
 	}
-	
+	*/
 	// Will not delete directory, just fail when trying to...
 	public void cleanCache(){
 		try{
@@ -110,28 +101,29 @@ public class ImageCache implements Runnable {
 					dis.close();
 					bis_info.close();
 				}catch(NullPointerException e){
-					Log.w("dk.nindroid.rss.ImageCache", "Unexpected null pointer exception caught.", e);
+					Log.w("Floating Image", "Unexpected null pointer exception caught.", e);
 					files[i].delete();
 					files[i] = null;
 				}
 			}
-			mExploreFiles.clear();
+			mFiles.clear();
 			int entries = Math.min(limit, files.length);
 			for(int i = 0; i < entries; ++i){
 				addFile(files[i]);
 			}
 		}catch(IOException e){
-			Log.w("dk.nindroid.rss.ImageCache", "Error removing old images...", e);
+			Log.w("Floating Image", "Error removing old images...", e);
 		}
 	}
 	
-	public void saveExploreImage(ImageReference ir){
+	public void saveImage(ImageReference ir){
 		if(ir == null) return;
 		String name = ir.getID();
+		if(exists(name)) return;
 		try {			
 			File f = new File(mExplore.getPath() + "/" + name + ".jpg");
 			FileOutputStream fos = new FileOutputStream(f);
-			ir.getBitmap().compress(CompressFormat.JPEG, 75, fos);
+			ir.getBitmap().compress(CompressFormat.JPEG, 85, fos);
 			fos.flush();
 			fos.close();
 			File f_info = new File(mExploreInfo.getPath() + "/" + name + ".info");
@@ -139,24 +131,24 @@ public class ImageCache implements Runnable {
 			fos_info.write((f.getAbsolutePath() + "\n" + ir.getInfo()).getBytes());
 			fos_info.flush();
 			fos_info.close();
-			synchronized(mExploreFiles){
+			synchronized(mFiles){
 				addFile(f_info);
-				mCached.put(f_info.getName(), mExploreFiles.size() - 1);
+				mCached.put(f_info.getName(), mFiles.size() - 1);
 			}
 		} catch (FileNotFoundException e) {
-			Log.w("dk.nindroid.rss.ImageCache", "Image could not be cached", e);
+			Log.w("Floating Image", "Image could not be cached", e);
 			return;
 		} catch (IOException e) {
-			Log.w("dk.nindroid.rss.ImageCache", "Image could not be cached", e);
+			Log.w("Floating Image", "Image could not be cached", e);
 			return;
 		}
 	}
 	
 	protected void addFile(File file){
-		mExploreFiles.add(file);
-		mCached.put(file.getName(), mExploreFiles.size() - 1);
+		mFiles.add(file);
+		mCached.put(file.getName(), mFiles.size() - 1);
 	}
-	
+	/*
 	public ImageReference getRandomExplore(){
 		if(mExploreFiles.size() == 0 || (!mActive || !Settings.useCache)){
 			try {
@@ -170,11 +162,11 @@ public class ImageCache implements Runnable {
 		Log.v("Image cache", "Adding random cached image");
 		return getFile(mRand.nextInt(mExploreFiles.size()));
 	}
-	
+	*/
 	public ImageReference getFile(int index){
 		InputStream is;
 		try {
-			File f_info = mExploreFiles.get(index);
+			File f_info = mFiles.get(index);
 			InputStream is_info = new FileInputStream(f_info);
 			BufferedInputStream bis_info = new BufferedInputStream(is_info, 256);
 			DataInputStream dis = new DataInputStream(bis_info);
@@ -183,7 +175,8 @@ public class ImageCache implements Runnable {
 				dis.close();
 				bis_info.close();
 				f_info.delete();
-				return getRandomExplore();
+				//return getRandomExplore();
+				return null;
 			}
 			// Read bitmap
 			File bmpFile = new File(bmpName);
@@ -212,8 +205,8 @@ public class ImageCache implements Runnable {
 	}
 	
 	public boolean exists(String name){
-		if(mExploreFiles == null) return false;
-		synchronized(mExploreFiles){
+		if(mFiles == null) return false;
+		synchronized(mFiles){
 			if(mCached.containsKey(name + ".info")){
 				return true;
 			}
@@ -222,7 +215,6 @@ public class ImageCache implements Runnable {
 	}
 	
 	public void addImage(String name){
-		Log.v("Image cache", "Adding cached image by request");
 		bank.addNewBitmap(getFile(mCached.get(name + ".info")));
 	}
 }
